Documentation build system work
Moved more code from APIMethod to APIEntry so that properties and accessors can take advantage of it; in this case the focus was the HTML generation code for fields that exist on APIEntry.

We now load and write out HTML for properties.

Fixed related opcodes for player.force_into_vehicle.