Documentation build system work
The pages for specific team and loadout settings are now more easily accessible.

Opcodes with two names now list both names in the sidebar.

Added CSS for a table of contents. Added a code snippets page which has a table of contents.

Blurbs, descriptions, and notes are now guaranteed to be wrapped in paragraph tags.