Documentation build system work
All loaded APITypes are now stored in an APIRegistry singleton, which allows objects inside of one type to refer to another type. The primary use of this is so that APIMethods' arguments can use the blurbs of their types as a description.

When created through APIRegistry, APIType now remembers what path it was told it loaded from, and relies on that path for path-related operations.