Documentation: incidents, overloads
Wrote an XML page for the "incident" type, with information dug out of the game files. Apparently I'd forgotten to do this before. Come to think of it, I may need to do the same for the "sound" type...

Added an "announce game start" example to the code snippets article.

The documentation build system now supports function overloads, though they have to be explicitly tagged as such in the XML.

Added more documentation comments to the build system's source code.