Documentation: sidebar improvements; namespace fixes
Fixed page title generation for namespace members.

The sidebar now auto-collapses folders on load, when appropriate.

Added a page for the global namespace. Modified pages for all variable types to specify the number of global variables that those types have.

Added an index page to the API reference. It doesn't have an auto-generated list of all functions in the language (no decent way to do that right now) but it'll do.