Map variant research (JS prototyping)
Fixed a mistake made when reading the code for loading Forge labels, which was causing us to load eight times as much data as we should.

Fixed an oversight for Forge labels: if the label count is zero, then we don't load any other data for labels (e.g. buffer sizes, etc.).

Decoded some more fields for Forge objects.