Code cleanup
Several flags-masks for Custom Games weren't actually defined in data, with the UI just using bare constants; this has been fixed.

Cut down on the number of preprocessor macros used to reduce copying and pasting in UI code. Changed the variable names that some of them use, in order to better reflect the fact that they are general-purpose.