Firefight UI work; team color fixes
Started work on adding Firefight round settings to the UI. Currently, we can display the states of all skulls for a round, as well as the dropship and order flags and squad counts for a round's waves. UI for squad types is a work in progress.

Fixed the code for editing teams' tertiary colors: those are ARGB instead of RGB, and we should now be handling the alpha component properly.