Firefight UI work
Custom skulls and base wave traits are now configurable.

Modified editor_state.h so that it doesn't bring in tons of game variant data headers, with individual UI components including the headers they need; this means that for the most part, changes to game variant data will be less likely to force us to recompile 99% of the UI. Unfortunately, the benefit from this is going to be limited, because so much of the UI runs directly through game variants rather than running through specific data structures; ReachEditorState::variantAbandoned is a particular offender.