Added support for Matchmaking Firefight variants
Matchmaking Firefight variants use a "gvar" block instead of an "mpvr" block. Brief inspection in a hex editor shows that the two blocks are identical except for the former's lack of an SHA-1 hash and associated data. ReachVariantTool has been modified to react accordingly.

ReachVariantTool also now warns on the absence of a "chdr" block, instead of failing with an error. Matchmaking Firefight variants use the yet-to-be-decoded "athr" block instead of "chdr", so this change makes them usable in ReachvariantTool, albeit with a bit of an annoying warning.