ATHR support; moved Player Species in UI
We now load ATHR blocks. Moreover, if a file is missing its CHDR block, we now properly generate one for it if possible, to be used when saving.

Moved the "Player Species" option from the UI's team settings to general settings. (This should've been in a prior commit; not sure what happened.)

Fixed label-buddies and tab order for the Firefight wave traits page.