Map variant research
Trying to decode Forge object rotations. Looks like the axis-angle "angle" value is loaded but not retained; instead, it's used on the spot to convert the axis-angle rotation to... probably an Euler rotation. Having some trouble recreating the process, though, since it's all float-based in the program and JavaScript uses doubles.