Docs: various fixes to generation; some fixes to content
Fixed entry-type constants on accessors, properties, and namespace members.

Removed the old `html_serializer` class since it's no longer used. This also allowed removal of the old Megalo syntax highlighter that it bundled.

Fully wired the category system up, and started using categories in the documentation.

Fixed generation of the "See also" sections.

We now properly encode files as UTF-8 when saving them.

We now load XML files differently: the QDomDocument::setContent API doesn't allow us to intercept and tamper with data at parse time, which is needed in order to add seamless support for HTML entities. I've adapted QXmlStreamReader to produce a QDomDocument, and created a custom entity resolver that it can handle. Said entity resolver currently only supports &mdash; and logs anything unsupported; as it happens, I don't use any other HTML entities at this time. We can add more in the future, or I can find a way to auto-generate a list for it to use somehow.

The object type list now uses a custom entity in its DTD for the "unused type" text, which means that that'll now be selectable (it wouldn't have been, with CSS).

Fixed quote styling in the sound list; we now use XML entities instead of bare characters, which protects us from encoding mishaps.