Work on v2.1.10
Added a menu bar to the script editor window; you can now save the file and access the help menu from there. To facilitate this, a lot of file I/O and file path code was moved off of the main window and onto ReachEditorState.

ReachVariantTool no longer mistakes Campaign game variants for Firefight game variants.

Added missing names to one of the Megalo "icon" enums.

The script compiler now validates format string tokens, emitting warnings or errors as appropriate. Documentation for format string tokens is now more thorough as well.

Fixed a bug in the script compiler that prevented it from reporting detailed error information for arguments to functions in the "game" namespace; a defect in the function it used to handle overload resolution in that namespace meant that it was treating every function as an overload of itself. (Overloaded functions cannot report detailed error information, because we don't know which overload the script author intended to use.)

Fixed a crash that occurred when compiling a script which incorrectly used a function with no return value in an assignment statement.

Fixed a mistake in the compiler's error handling which caused it to potentially omit a function call from its compiled output if one or more of the arguments to that function generated a warning.

Added documentation pages for player traits. Fixed documentation page titles for indexed lists (e.g. script_widgets): the VAR tags are no longer written to the TITLE tag.

Added a documentation article for biped positions/rotations.

When saving a file from the main window, the main window's title is temporarily changed to "File saved!" Hopefully at some point in the future, I'll think of a better way to alert the user to a Ctrl+S press being registered.

The main window now uses Qt's "standard" key combinations for Open, Save, Save As, and Help.