v2.1.10
Fixed the compiler's handling of for-loops that have been set as event handlers: we now wrap the loop in a do/end block and use the wrapper as the event listener. This is needed because the game will not consider a file valid if a for-loop is an event handler.

The options window should now consistently open to its first tab, regardless of which one I last edited in Qt Designer.

Fixed how the documentation generator produces sidebar links, when handling opcodes with multiple names. There should no longer be 404s for the secondary names.

The documentation generator now allows you to append extra text to an API type's "scope" information, by placing text content inside the "scope" element.

The documentation generator now properly strips HTML out of page titles when writing them into a TITLE element.

Corrected and improved documentation for object.set_waypoint_range.

Documentation now mentions that only 256 objects at a time can have member variables.