Reflex enum improvements
For the refactored cobb::reflex::enumeration, removed lots of code duplication regarding how we handle underlying values, and implemented the ability to cast an enum value to a string. The secret to cutting down on code duplication is that we now generate an array of structs that map underlying values to type indices and sub-indices.

Fixed the constraint on members' explicit values being representable in the chosen underlying type.