Continuing work on reimplementing Megalo operands and opcodes
Defined more Megalo operands. Modified `base_enum_operand` to use the new-style reflection-capable enums.

A difference from the original code is that relative-member operands and variable-and-member operands are now templated on the types of the basis value and member value. There's no specific need for this -- the only relative-member operand recognized by the game is context_object.timer, and the only variable-and-member operand recognized by the game is object.player -- but it feels like a cleaner way to specify exactly what these operands are.

Added a `cobb::reflex::is_enumeration` concept.