Work continues on Megalo data serialization
Created `bitfloat` to complement `bitnumber`, for use with the TU1 floats.

All operands have classes now.

Converted the operand types for `player` and `team` to subclasses, so we can further configure the typeinfo (i.e. for static counts).

Wrote classes for TU1 data and trigger entry points, and read functions for opcodes.