Work on Megalo operands; reflex massive_enumerations
Created `cobb::reflex::massive_enumeration` to account for the fact that MSVC will gorge itself on dozens of GB of RAM until it chokes to death, if a normal reflex enumeration has a few hundred items in it. The "incident" types in Megalo were the main offender here.

Updated Megalo operand types to use the new-style reflex flags masks.

Various compiler fixes.