Bitnumber no longer tries to automatically extend sign bits
Attempted to build a natvis file for reflex flags masks; unfortunately, the natvis engine can't match the type properly (i.e. it can't tell that any given variable has the type that the natvis file is meant to refer to). I think it's choking on the template parameters.

Modified `halo::bitnumber` so that sign bit extension is something you have to opt into. This makes it much more obvious when it is and isn't happening, making maintenance and debugging easier. That said, the `bytenumber` shortcut template opts you in if you use a signed non-enum type. Additionally, `bitnumber::read_bits` never extends the sign bit, regardless of the type you template it on; you can manually run `cobb::apply_sign_bit` afterward if you must.