Load process messages now hold the bitstream position they're logged from
All load process errors that occur when bitreading are now logged through bitreader. Comparable changes still need to be made to bytereader.

Load process messages logged through bitreader will now include either the reader's stream position at the time of the message, or (if the caller specifies) a custom position. The current pattern is for callers to manually "rewind" the current stream position to where the bad value was read from.