Compiler error fixes; SHA-1 changes
Changed how we generate MPVR SHA-1 hashes. It might still be wrong, but we're closer. The new SHA-1 code should be correct, but I'm not sure I'm using it on the correct set of bytes yet.

Fixed bitwriter support for operands. Implemented it for string tables, though we don't have the fallbacks for too-large-tables that the release build has yet.

With these changes, the program now builds!