More save fixes
Fixed serialization of map permissions, and of floats. Modified trigger serialization in an attempt to be more Bungie-consistent.

Given current settings, we are now consistent with assault_054.bin up to offset bitstream+0x21E5. We go wrong somewhere in the Megalo conditions, which start at offset bistream+0x21B1+0b. We may be writing the conditions properly but not in a Bungie-consistent order, or we may be writing a condition improperly; hard to know.