Further fixes to resaving
Fixed the issue with conditions not matching on resave. It was a combination of problems which interfered with identity comparisons (i.e. `&a == &b`). The first was that the opcode function lists needed to be marked inline constexpr (so that all access to them uses the same instances i.e. the same memory addresses). The second was that even with them marked as inline constexpr, MSVC apparently doesn't know what references are, so constexpr references to inline constexpr variables end up compiling as non-references (i.e. separate copies, with separate memory addresses). Constexpr pointers to inline constexpr variables work.

Rewrote the logic for `trigger::extract_nested_trigger_indices` to be more self-contained. It's functionally equivalent (pointer/reference compiler bugs aside) but a bit cleaner now.

With these changes, and with the string table optimization set to match Bungie, the built-in `assault_054.bin` file now resaves consistently with the original up to file byte offset 0x29F1 (Megalo bitstream byte offset 0x26D9), which is a decent way into Megalo action data (so we probably got an operand or few wrong).