Work continues on the token scanner
The scanner now uses constexpr lists of all "syntax characters" divided by meaning and handler behavior. This in turn allows us to automatically generate a list of "syntax characters" for use when parsing for "words," which in turn means words don't have to be limited to ASCII or some other easily-testable character set. Want aliases in foreign-language character sets? No problem!

When scanning in number literals, we no longer treat the sign as part of the literal. Given the input "-0x123", this is no longer seen as a single number; it is now seen as a unary negation operator applied to the number literal "0x123". This makes scanning much easier to implement, and allows the unary operators "-" and "+" to be treated the same as any other operator.