Bolt: no more tagged string literals; scanner now throws
Renamed AST content (folder and namespace hierarchy) to reflect the possibility of adding a new language. The current language is Bolt; at some point in the future, I hope to work on a language with symbolic variables and register allocation, which would be called Rivet.

Removed tagged string literal syntax (`"foo"#5`), as it doesn't accomplish the use case I intended it for. Instead, I'll have a hardcoded collection `script_strings[n]`, and aliases will be able to target items in this collection.

The scanner now throws an exception upon encountering an unterminated string literal. Additionally, more diagnostic information is collected.

Fleshed out a bit more of the parser. At this point, the scanner feels like it's mostly complete (extracting pragma directives from inside of comments is a larger challenge; the current design is not well-suited for this), so we should start moving on to the parser which will consume the scanned tokens.