Revised pragma syntax; implementation pending
Revised the syntax for pragmas; instead of placing them in comments, we're just going to make them a normal keyword. There's no point in maintaining backward-compatibility when adding language features; things like "for each object of type frag_grenade" will already fail to compile in Bolt v1. We can now remove all of the comment-scanning code, which will greatly simplify scanning.

Corrected an error in the Bolt v1 grammar -- a copy-and-paste mistake for "for" loops resulting in v2-only syntax ("for each object of type") being listed as supported.