Bolt: Scanner testing and fixing
The scanner is now undergoing testing.

Created natvis files for scanner tokens and related constructs.

Deleted `token_type::operator_assign_bare`, as it was both wrong and invalid (it's impossible for the scanner to tell whether any given '=' glyph is an assignment operator or some other construct; examining context and figuring out how the glyph is being used is the parser's job).

FIX: The bare-assign operator `=` is no longer mistaken for the equality comparison operator `==`. The issue is that when handling multi-character operators, we didn't advance past the first glyph after confirming its identity, so we saw it twice. In fact, this also would've made it impossible to properly match any multi-character operator.

FIX: Removed the null glyph passed to character scan functors, as it was being mistaken for an identifier.

FIX: When the scanner extracts a word (as part of a nested character scan), the inner scan (for extracting the word) no longer causes the outer scan (for scanning the source in general) to miss the glyph after the end of the word.

FIX: Number extraction had the same issue as word extraction; this is now fixed.

BUG: The above two fixes suggest that the way we're actually scanning through the source is clumsy and needs refinement. Having nested character scans return `stop_before` is a hack, and one that will break if multiple nested character scans are run in a row (e.g. extracting multiple words at a time, if for some reason we wanted to do that). We need a better solution.

BUG: The scanner assumes that every '.' glyph is a number literal, breaking the scanning of strings like "current_object.number[3]".

BUG: The above bug hints at another issue: it is impossible to tell an uninitialized `literal_data_number` apart from an invalid one (e.g. `0b011012345`).

BUG: Remove `token::pos`; because `token` subclasses `item_base`, it already has start and end positions, which we fill in.