phrase_tree is a failure and will be deleted
Changed `cobb::cs::value` to an array (as opposed to a `std::array`) since we never iterate over it. I did this in the hopes that IntelliSense might display compile-time strings more sensibly this way. It does not.

`phrase_tree` is a failure and will be deleted. Working out the logic this way is simply not worth the effort.