Bolt: identifier member access constraints; work on declarations
User-defined function names that contain member access (e.g. `function a.b.c[3]() end`) are now a parser error.

Wrote code for parsing alias declarations, variable declarations, and user-defined enum declarations. The latter is unfinished; the former two are untested.

When parsing a block, we assert that the block is properly closed. This assertion originally used a helper analogous to std::scope_exit; it now uses a helper analogous to std::scope_success, such that the assertion isn't made if we exit the function due to a thrown exception. This is because if a parse exception is thrown while parsing the block itself or its contents, the block will not be closed. (That, if anything, is helpful, as it means that `parser::current_block` directly tells us the block hierarchy at the error site.)

Megalo: Moved the variable network priority enum out to its own file, and to an identifier that's easier to grab from afar (e.g. for the Bolt AST).