Reflex enum experiments: Mk. III
Thought of a way to create reflex enums that avoids the hellish classnames we see with the current approach.

The idea is that you declare a class; then, you specialize a "member list" template on your declared class to define your enum members; then, you finally define your class, using CRTP to inherit from a base "reflex enum" class. The use of the separate template for member definitions allows us to sidestep the limitations of CRTP -- namely, that the base class's definition (and the definitions of its members) cannot be dependent on the members of the derived class, as the derived class "doesn't exist yet" at the time CRTP is used to specialize the templated base.

The current draft at least works in IntelliSense, and I want to extend it to support "gap," "range," and "nested enum" members as the previous attempt at reflex enums does.