Bolt: Started work on built-in identifiers
All attempts at redesigning reflex enums have been scrapped ("Exclude from Project") for now. Reflex3 is the closest we got to success, but is blocked by what appears to be incomplete P1787 support in MSVC. This results in IntelliSense failing on nested enums despite the compiler allowing them, and it's the reason we have to globally-namespace the enumeration traits template to allow specialization in the first place.

This commit includes a scrapped reflex4 attempt that would've implemented the "member lists" for enumerations as values in themselves. This is non-viable: template metaprogramming is significantly more constrained for values.

Implemented helper templates for constexpr tree structures.

Started work on built-in identifiers in the new Bolt parser: defined a constexpr tree of identifier names intended to serve as a representative sample for the identifiers built into Bolt. This includes multiple identical identifiers distinguishable only by type (but we have not yet implemented a type system): multiple opcode-specific built-in enums have members with conflicting names (e.g. "none"), and the only way to handle these consistently with Bolt v1 is to allow an identifier in script to resolve to any of a set of identical built-ins, with us selecting the one to use based on the type that was expected where the identifier appeared.