Support for temporary variables; numerous other improvements
Added `local_player` and `local_team` as alternate names for `hud_player` and `hud_player_team`. The former two names are used in MegaloEdit, and they lead me to believe that these variables may(!!!) be used for more than just format strings.

Compiler: added `VariableReference::is_transient` and `VariableReference::is_usable_in_pregame` based on reverse-engineering of MegaloEdit and its error messages. Neither is used yet. The latter is going to be unwieldy to implement because the compiler compiles statements into opcodes as they're parsed (with blocks compiled into triggers at the end of the script code); this makes it difficult to review already-compiled material, which we'd have to do in order to validate functions called by the pre-game trigger.

Added the following alternate opcode names for parity with MegaloEdit:
 - set_round_card_title -> set_objective_text
 - set_round_card_text -> set_objective_allegiance_name
 - set_round_card_icon -> set_objective_allegiance_icon
 - get_button_press_duration -> get_button_time

Renamed `player.unk09` to `player.money` to match MegaloEdit.

Updated documentation to match the above changes; additionally, documented `object.create_object` as an alternate name for `object.place_at_me`.

Documented length limits on displayed strings for HUD widgets. Documented all new variant string IDs.

Updated the backend and UI to properly display and handle the team scoring type, after getting the values out of MegaloEdit.

Capped powerup durations to 120 for consistency with MegaloEdit's validation.

Made common-sense changes to get rid of the need for macros in VariableDeclarationSet. I don't have time to fix the rest of RVT's cringe code; it'd be better and easier to rewrite the whole backend from scratch (again, since the last attempt broke when compiler updates killed my reflection-capable enum shenanigans).

---

Added support for MCC's new temporary variables to the loader and compiler. These changes included...

Variables are implemented as a type (e.g. `e_custom_reference_type`) followed by potentially arbitrary data; MegaloEdit uses a switch-case. We abstract this into a "scope" type, a "which" type, and an "index" type, where the former determines which of the latter two are present and what they mean. Temporary numbers are implemented as a scope-and-index pair; all other temporaries are implemented as scope-and-which pairs (i.e. one "which" value per variable index, similarly to globals).

Accordingly, several new VariableScopeWhichValues needed to be added, along with a new VariableScope. We didn't need to add a new declaration; temporaries do not need to be declared.

The compiler needed a new namespace for temporaries. Several pieces of code in the compiler, and in the compiler- and decompiler-related parts of variable opcode functions, assume that non-member variables are globals, and these needed to be updated.

The compiler needed a warning message for if you declare a temporary; it's an error if you specify an initial value or networking priority.

Opcode arg typeinfo needed to be able to store the max temporary count, and a pointer to the first "which" value that represents a temporary for the given type. On reflection, the former is probably unnecessary; I believe the `VariableReference` can (and does) check the MegaloScope's maximums instead.