Variable allocation now banned in functions; started VariableReference refactor
The compiler no longer allows `alias <name> = allocate temporary <type>` inside of user-defined functions, as there's no way to reliably identify available variable indices for use there.

Wrote a version of `AccessorRegistry::get_variably_named_accessor` which doesn't require access to a `Compiler`. Should remove the old version (which does) once I'm sure this works reliably.

Started work on a feature which is intended to warn about user mistakes surrounding variables allocated to aliases. The goal is that if a variable is currently aliased via `allocate` syntax, and you create another alias pointing directly to that variable (i.e. not an alias pointing to the allocated alias), then we should emit a warning.

Started work on a rewrite of `Script::VariableReference`, because it... sucks, frankly. It's difficult to look at that class and answer questions like "How do I tell if this refers to `global.number[0]`?", and a lot of the code surrounding it is a veritable rat's nest. The rewrite attempts to simplify things by using std::variant, so in effect, a `VariableReference` only has those fields which correspond to whatever variable it actually refers to. It's more explicit about the meaning of its own contents, rather than containing a big soup of fields that have different context-dependent meanings where each field's value influences the meaning of other fields in ways that aren't clear at a glance.

There are limits to how clean we can make this system because ultimately, the whole type/scope/which/index system for variables is not fully reflective of the underlying reality of what variables in Megalo mean and how they're stored. It would've been more accurate to do something like:

   #pragma pack(push)
   #pragma pack(1)
   struct number_reference { // timers are the same, except no union+constant
      enum class type : uint8_t {
         // ...
      };

      enum type type;
      union {
         int16_t constant;
         struct {
            uint8_t context_index; // W in player[W].number[I]
            uint8_t target_index; // I in player[W].number[I] or global.number[I]
         } fully_qualified_variable;
      };
   };

   struct object_reference { // players and teams are the same
      enum class type : uint8_t {
         // ...
      };

      enum type type;
      struct {
         uint8_t top_level_index; // W in player[W].object[I] or global.object[W]
         uint8_t member_index; // I in player[W].object[I]
      } fully_qualified_variable;
   };
   #pragma pack(pop)

(We don't actually need to pack the stuff; that's just how it works in Reach's game engine under the hood. The point is that this is both clearer and more accurate than "scope, which, and index." It doesn't need weird edge-cases for integer constants, and it enforces a clearer separation between how numbers and timers are handled versus how objects, players, and teams are handled -- a separation that would be easier to metaprogram.)