Reverted VariableReference changes; we now warn on allocate/explicit alias collisions
Rolled back all major changes to VariableReference.

The compiler now warns if you set two aliases to the same variable, with an earlier alias using `allocate` syntax and a newer alias using an explicit reference to the variable that happened to be chosen by `allocate`. (In simpler terms: we now warn on accidental collisions.)

Wrote some experimental code for trying to redesign how variable references are handled. Disabled all of it. No good way to integrate more modern code into a codebase this ancient; better to just rewrite it all someday.