Allocation-to-alias now allowed, and safe, inside of user-defined functions
Implemented a workaround that makes it possible for us to safely offer allocation-to-aliases from inside of user-defined functions. We just keep track of all of the aliases that exist within each in-scope user-defined function, and when allocating-to-alias and looking for a non-aliased variable, we check against those aliases as well.

This required a change to how the compiler handles top-level blocks. Normally, once we reach the end of a top-level block, we compile it and then delete it. However, if the top-level block is a user-defined function, then this deletes its aliases, leaving dangling alias pointers in the UserDefinedFunction object (which represents an in-scope function). The quick-and-dirty workaround is: instead of deleting top-level blocks after compiling them, we now just move them to a separate list for safekeeping, and delete them later.

---

As part of this change, the code for `allocate <base>.<type>` and `allocate temporary <type>` has been unified, for easier maintenance. A helper object, `Compiler::_allocation_request_type`, has been added in order to make the code just a little bit actually organized.

---

Documentation updates:

The article for user-defined functions now explains that functions are block-scoped.

Updated the documentation for allocate-to-alias, to remove remarks about the old limits and to add details about the nuances of allocation-to-aliases inside of user-defined functions. It's more complicated than I'd like, but that's what I get for writing an assembler with Lua vibes and telling everyone it's a "compiler," haha.