Temporary variables are now initialized
The compiler will now auto-initialize allocated temporary variables when they are first used, unless that usage already fully overwrites the temporary variable's value. This is more-or-less consistent with MegaloEdit's handling of temporaries.

To that end:

We have a new struct, `variable_usage_set`, which is essentially a flags-mask with one flag for every global or temporary variable. OpcodeArgValue now has a virtual member function which will set the flags for any variables used in the operand. (Remember that some operands include variable references as part of their contents, e.g. format strings.)

The compiler maintains one `variable_usage_set` to track all allocated temporary variables, and another to track which ones have actually been initialized. After we compile and insert a Script::Statement, we pass it to a function which checks whether it uses any temporary variables and, if so, whether those temporaries are allocated, uninitialized, and not fully overwritten by the statement itself. If so, we compile an assignment to zero or none, and insert that before the statement.

The code for `variable_usage_set` is messy. I could've tried to metaprogram it, but since so little of ReachVariantTool's existing internals are metaprogrammed -- particularly the variable scope definitions, with variable count limits and whatnot -- metaprogramming would've taken vastly more effort than it would've saved. I desperately want to rewrite ReachVariantTool from scratch. Maybe someday...