Round card preview
The localized string editor now offers a round card preview, which opens as a non-modal dialog.

This is not perfect. The preview shows the last textbox you've edited since opening it, which means it opens blank; we should default it to the first non-blank string (typically English). The text width is slightly off despite my best efforts; I suspect it's a combination of ClearType on Qt's end, and inconsistent and possibly inaccurate glyph rendering in-game.