Can now load *.mglo files; refinemenets to round card preview
Round card preview: experimented with a hack to forcibly disable ClearType, gated behind a constexpr bool, to see if it'd improve accuracy. It did not. Settled on 99%-width text.

Round card preview: when initially opened, it now shows the first non-blank localization in the current string, if any.

ReachVariantTool can now load raw MGLO files. It's simple enough; they're just all of the Megalo-specific data in the MPVR block.