Bugfixes: temporary vars; MGLO resaving; Forge label on-demand creation
Fixed a bug where references to nested members of temporary variables were compiling as nested members of global variables.

Fixed a bug where loading a MGLO file and resaving it would write out BIN data: we weren't setting the MPVR type on load.

When Forge labels are created on demand, we now only emit a notice for the first time the new label is referenced.

Forge labels will now be created on demand for operands, too, not just blocks. This also fixes an issue where if your script had a for-loop for a non-existent label, but also had operands referring to that same label name, the latter would fail to compile.