We no longer auto-initialize temporaries, as it breaks user-defined functions
Scrapped the feature to default-initialize temporary variables if you try to modify or read them before setting them. We were getting false-positives for uninitialized access, in situations where you define a function that uses a temporary from an outer scope, and the function is defined before the temporary is assigned, but the temporary is always assigned before the function is called. We were compiling "initializations" into the function which clobbered the variable's value at call-time. There is no way to avoid this problem in a single-pass compiler.