Documentation updates
Updated the documentation for `object.place_at_me` to better describe the rotation of the created object.

Corrected the descriptive text for `object.set_pickup_priority`.

Reverse-engineered `object.get_distance_to` and documented that it rounds up when dealing with existent objects, and returns zero when dealing with non-existent objects.

Documented the limit on how many objects can have Megalo data attached. (The internal game state object for this used to be called `megalo objects`; it seems like game state singletons no longer have names embedded in the executable.) This is based on reverse-engineering the getter for scripted object data, which takes a bool and can create that data.