Inno Setup, uh, setup, and related fixes and build config
Qt VS Tools has a bug which prevents windeployqt from properly copying the VC redistributable into the program's output folder. To work around this, I've written a custom build task with C# code which grabs that file.

I've moved RVT's build output to a subfolder to isolate it from the output folders for the documentation processors.

Added a project with Inno Setup code files, for use in generating an installer for RVT. The installer will run the VC redistributable if present and if more recent than whatever the end user already has installed.