Fixed player.money setter; object type research; documentation
Action 38, previously identified as `player.increment_req_money_by`, was incorrect. In reality, it's a setter, matching the `player.money` getter and taking a math operator.

Updated documentation and in-program metadata for several Megalo-spawnable object types.

Documentation: added a page for Map and Game Settings as they exist in Firefight.

Documentation: updated my description of how triggers are stored under the hood. The optimization described was suggested by a user, but is unfortunately not practical in RVT given RVT's implementation. If I ever have time to refactor RVT as thoroughly as I would like (read: rewrite the whole damn thing), this is one of the things I should give particular focus to.